move some messages to the constant
