+++ build.gradle
@@ -12,7 +12,7 @@ repositories {
 
 dependencies {
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-test', version: '2.3.2.RELEASE'
-    compile  group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '2.3.2.RELEASE'
+    compile group: 'org.springframework.boot', name: 'spring-boot-starter-web', version: '2.3.2.RELEASE'
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-parent', version: '2.3.2.RELEASE', ext: 'pom'
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-jdbc', version: '2.3.2.RELEASE'
     compile group: 'org.mybatis.spring.boot', name: 'mybatis-spring-boot-starter', version: '2.1.3'
@@ -21,11 +21,10 @@ dependencies {
     compile group: 'org.springframework.boot', name: 'spring-boot-starter-security', version: '2.3.2.RELEASE'
     compile group: 'org.apache.commons', name: 'commons-lang3', version: '3.11'
     compile group: 'org.hibernate', name: 'hibernate-validator', version: '6.1.5.Final'
+    compile group: 'org.flywaydb', name: 'flyway-core', version: '6.5.3'
     providedCompile group: 'org.projectlombok', name: 'lombok', version: '1.18.12'
-
     testCompile group: 'org.mockito', name: 'mockito-core', version: '2.25.0'
     testCompile group: 'org.jmockit', name: 'jmockit', version: '1.9'
-    compile group: 'org.flywaydb', name: 'flyway-core', version: '6.5.3'
     testCompile group: 'junit', name: 'junit', version: '4.12'
     testCompile group: 'org.mybatis.spring.boot', name: 'mybatis-spring-boot-starter-test', version: '2.1.3'
 }
