+++ src/main/java/com/kingland/bootcamp/common/consts/MessageConst.java
@@ -1,0 +1,18 @@
+/**
+ * Copyright 2020 Kingland Systems Corporation. All Rights Reserved.
+ */
+package com.kingland.bootcamp.common.consts;
+
+/**
+ * Message constants
+ *
+ * @author KSC
+ */
+public class MessageConst {
+
+    public static final String USER_NOT_FOUND = "Could not find user";
+    public static final String LENGTH_NOT_VAILD = "The length not valid";
+    public static final String USERNAME_NOT_VAILD = "Only letters numbers and underscores are allowed";
+    public static final String PASSWORD_NOT_VAILD = "At least one uppercase letter, one lowercase letter, one number and one special character";
+    public static final String DO_NOT_HAVE_ACCESS = "Accessed resources require identity authentication";
+}
