+++ src/main/java/com/kingland/bootcamp/common/consts/SecurityConst.java
@@ -1,17 +1,19 @@
 /**
  * Copyright 2020 Kingland Systems Corporation. All Rights Reserved.
  */
-
-package com.kingland.bootcamp.common;
+package com.kingland.bootcamp.common.consts;
 
 /**
+ * Security constants
+ *
  * @author KSC
  */
 public class SecurityConst {
 
     /* Login Jump URL*/
     public static final String AUTH_REQUIRE = "/authentication/require";
-
     /* Form login URL*/
     public static final String AUTH_FORM = "/authentication/form";
+    /* Login URL*/
+    public static final String AUTH_Login = "/login.html";
 }
