--- src/main/resources/message/messages.properties
@@ -1,12 +1,0 @@
-not.null=* This option is required!
-user.not.exists=User does not exist/password error!
-user.password.not.match=User does not exist/password error!
-user.logout.success=Logout successfully!
-length.not.valid=The length must be between {min} and {max} characters!
-length.less.valid=The length must be less than {max} characters!
-user.login.success=Login successfully!
-user.unknown.error=Unknown error!
-no.view.permission=You have no permission!
-user.does.exists=Username already exists!
-operate.success=operate successfully!
-operate.failed=operate failed!
